pyproject.toml
[project]
name = "sonic-bloom"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "rich>=13.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0",
    "python-dotenv>=1.0",
    "pyobjc-core>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-ScriptingBridge>=10.0",
    "pyobjc-framework-MediaPlayer>=10.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
all = ["openai>=1.0"]

[project.scripts]
sonic-bloom = "sonic_bloom.__main__:main"
